// src/server/api/models/banking/models/LoanAccountPayment.ts
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

import type { LoanAccountPaymentStatus } from './LoanAccountPaymentStatus';
import type { LoanAccountSchedule } from './LoanAccountSchedule';
import type { LocalDate } from './LocalDate';
import type { LocalDateTime } from './LocalDateTime';
import type { SubAccountType } from './SubAccountType';
import type { UUID } from './UUID';

export type LoanAccountPayment = {
    /**
     * The unique ID for the record. This is the ID that is used in all services and related records when referencing the record.
     */
    id: UUID;
    /**
     * The ID of the user that created the record.
     */
    createdBy: string;
    /**
     * When the record was created.
     */
    createdAt: LocalDateTime;
    /**
     * The ID of the user that last modified the record.
     */
    updatedBy: string;
    /**
     * When the record was last modified.
     */
    updatedAt: LocalDateTime;
    /**
     * Indicates if this is a deleted record
     */
    deleted: boolean;
    /**
     * The journal ID used to process the payment
     */
    journalId?: UUID | null;
    /**
     * The transaction notes
     */
    notes?: string | null;
    /**
     * The ID of the deposit account used to process the payment. This is optional but if specified, it has to be a deposit account ID that belongs to the client that owns the loan account.This deposit account has to be active, in the same currency as the loan and a have positive balance.
     */
    depositAccountId?: UUID | null;
    /**
     * The deposit account number used to process the disbursement
     */
    depositAccountNo?: string | null;
    /**
     * The deposit account label used to process the disbursement
     */
    depositAccountLabel?: string | null;
    /**
     * The loan schedule that this payment is for. This is optional but if specified, it has to be a schedule that belongs to the loan account.
     */
    loanAccountSchedule?: LoanAccountSchedule | null;
    /**
     * A reference code for the account record in an upstream system
     */
    refCode: string;
    /**
     * A unique payment code generated by the system
     */
    code: string;
    /**
     * The payment description
     */
    label: string;
    /**
     * The status of the payment
     */
    status: LoanAccountPaymentStatus;
    /**
     * The amount allocated to principal
     */
    principal: number;
    /**
     * The amount allocated to interest
     */
    interest: number;
    /**
     * The amount allocated to taxes.
     */
    taxes: number;
    /**
     * The amount allocated to penalty.
     */
    penalty: number;
    /**
     * The amount allocated to fees
     */
    fees: number;
    /**
     * The total amount received
     */
    paymentAmount: number;
    /**
     * The transaction currency
     */
    currency: string;
    /**
     * The value date for the transaction. This can be used to back date the transaction
     */
    paymentDate: LocalDate;
    loanAccountScheduleId?: UUID | null;
    productType: string;
    accountType: SubAccountType;
    productId: UUID;
    transactionSourceId: UUID;
    productLabel: string;
    productCode: string;
    clientId: UUID;
    clientLabel: string;
    clientCode: string;
    loanAccountId: UUID;
    loanAccountNo: string;
    loanAccountLabel: string;
    totalPrincipalAmount: number;
    interestRate: number;
    loanStartDate: LocalDate;
    expectedMaturity: LocalDate | null;
};